% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial-censor.R
\name{artificial_censor}
\alias{artificial_censor}
\title{Artificial Censoring}
\usage{
artificial_censor(x, quantiles, detlims)
}
\arguments{
\item{x}{numeric. Original data to be censored.}

\item{quantiles}{numeric. Censor by quantile(s) between 0 and 1.}

\item{detlims}{numeric. Censor at chosen value(s).}
}
\description{
Converts bare numeric into a left-censored quantity with less information.
}
\details{
Only one of the arguments \code{quantiles} and \code{detlims} should be supplied.
}
