% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{cens_ratio}
\alias{cens_ratio}
\title{Compute Degree of Censoring}
\usage{
cens_ratio(x, na.rm = FALSE)
}
\arguments{
\item{x}{object of class \code{\link{Nd}}.}

\item{na.rm}{logical. NA observations ignored when \code{TRUE}.}
}
\description{
Compute Degree of Censoring
}
\seealso{
Other \code{Nd} functions: 
\code{\link{impute.Nd}()},
\code{\link{is.Nd}()},
\code{\link{plot.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{\code{Nd} functions}
