% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-impute.R
\name{impute.Nd}
\alias{impute.Nd}
\alias{impute_by_factor}
\alias{impute_by_ros}
\title{Impute Non-Detect Data}
\usage{
\method{impute}{Nd}(x, method = "DL2", delog = FALSE, ...)

impute_by_factor(value, is_nd, impute_fct = 0.5)

impute_by_ros(value, is_nd)
}
\arguments{
\item{x}{object of class [`Nd`].}

\item{method}{coercible to function. See **Method** section.}

\item{delog}{logical. When `TRUE`, values are exponentiated before
imputation and log transformed afterwards.}

\item{...}{extra arguments passed to `method`.}

\item{value, is_nd}{numeric, logical. Used in pre-defined methods.}

\item{impute_fct}{numeric. Multiplicative factor, defaults to 0.5.}
}
\description{
Impute Non-Detect Data
}
\section{Functions}{
\itemize{
\item \code{impute_by_factor}: Replace each non-detect with the product of some factor (`impute_fct`)
and the detection limit.

\item \code{impute_by_ros}: Apply Regression on Order Statistics (ROS) method proposed by Lee and
Helsel in the `NADA` package.
}}

\section{Method}{

An imputation method can be pre-defined and identified via a character,
such as
  - `"DL2"`, half detection limit,
  - `"ROS"`, regression on order statistics via `NADA` package.

Other types are passed to [rlang::as_function()] and is expected to return
a numeric vector given arguments `value` and `is_nd`.
}

\seealso{
Other `Nd` functions: 
\code{\link{cens_ratio}()},
\code{\link{geom_Nd}()},
\code{\link{is.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{`Nd` functions}
