% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-impute.R
\name{impute}
\alias{impute}
\title{Impute Generic}
\usage{
impute(x, ...)
}
\arguments{
\item{x}{object used to choose a method.}

\item{...}{extra arguments used for extensibility.}
}
\value{
Regardless of input class, should always return same-length numeric.
}
\description{
Transform data from some class to a simplified numeric to be used in place.
}
