% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{is.Nd}
\alias{is.Nd}
\title{Check `Nd` Class}
\usage{
is.Nd(x)
}
\arguments{
\item{x}{object of class [`Nd`].}
}
\description{
Check `Nd` Class
}
\seealso{
Other `Nd` functions: 
\code{\link{cens_ratio}()},
\code{\link{geom_Nd}()},
\code{\link{impute.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{`Nd` functions}
