% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-visualise.R
\name{layer_Nd_line}
\alias{layer_Nd_line}
\title{Add \code{ggplot2} Layer for Non-Detect Data}
\usage{
layer_Nd_line(y, ..., scale = TRUE)
}
\arguments{
\item{y}{expression. Variable to use on y-axis.}

\item{...}{Extra arguments passed to \strong{both} geoms.}

\item{scale}{logical. When \code{TRUE}, a custom \code{scale_shape_manual} is added.}
}
\description{
In absence of a rigorous \code{ggplot2} extension, this function accepts a
mapping to \code{y} and returns \code{geom_point}, \code{geom_line} and optional scale.
}
\note{
The main use for \code{y} is to be placed in \code{aes(...)} with added
information such as \code{value} or \code{is_nd}. As such no data is handled here.
}
