% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-class.R
\name{Nd}
\alias{Nd}
\alias{Nd.default}
\alias{Nd.Nd}
\alias{Nd.numeric}
\alias{Nd.character}
\title{Non-Detect Data}
\usage{
Nd(x, ...)

\method{Nd}{default}(x, ...)

\method{Nd}{Nd}(x, ...)

\method{Nd}{numeric}(x, is_nd, ...)

\method{Nd}{character}(x, ..., na = c("", "NA"))
}
\arguments{
\item{x}{object. Input to be converted to `Nd`.}

\item{...}{These dots are for future extensions and must be empty.}

\item{is_nd}{logical. Required when `x` is numeric.}

\item{na}{character. Vector of values to be interpreted as NA.}
}
\value{
An object of class `c("Nd", "Surv")`.
}
\description{
S3 class that represents left-censored data commonly referred to as
  **non-detect** data.
}
\details{
The object is made up of a numerical component, `value`, that
  represents an observed concentration or level of detection depending on
  whether the logical component, `is_nd`, is `TRUE` or `FALSE` respectively.
}
\section{Survival}{

  This class is an extension to the [survival::Surv] class; a special case
  where left-censoring is implicitly assumed.
}

\examples{
  Nd(c(1.1, 0.5, 0.25), is_nd = c(FALSE, TRUE, TRUE))
  Nd(c("1.1", "<0.5", "ND<0.25", "1.312454"))

}
