% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Nd_example}
\alias{Nd_example}
\title{Helsel (1988) Example Data}
\format{
An object of class Nd (inherits from Surv) with 18 observations.
}
\source{
\url{https://doi.org/10.1029/WR024i012p01997}
}
\usage{
Nd_example
}
\description{
Used as toy example in Appendix B of Helsel and Cohn (1988).
}
\references{
Helsel, D.R. and Cohn, T.A., 1988. Estimation of descriptive
statistics for multiply censored water quality data. Water Resources
Research, 24(12), pp.1997-2004.
}
\keyword{datasets}
