% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-visualise.R
\name{plot.Nd}
\alias{plot.Nd}
\alias{autoplot.Nd}
\title{Plot Non-Detect Data}
\usage{
\method{plot}{Nd}(x, ..., type = "quantile")

autoplot.Nd(x, ..., type = "quantile")
}
\arguments{
\item{x}{object of class \code{\link{Nd}}.}

\item{...}{Extra arguments passed to other methods.}

\item{type}{choice. For future extensions where plot type is to be chosen.}
}
\description{
Visualisation methods for non-detect data.
}
\seealso{
Other \code{Nd} functions: 
\code{\link{cens_ratio}()},
\code{\link{impute.Nd}()},
\code{\link{is.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{\code{Nd} functions}
