% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-visualise.R
\name{quantile_plot_Nd}
\alias{quantile_plot_Nd}
\title{Quantile Plot of Non-Detect Data}
\usage{
quantile_plot_Nd(x)
}
\arguments{
\item{x}{object of class \code{\link{Nd}}.}
}
\description{
Quantile plot comparing the rank to the value of data.
}
\keyword{internal}
