% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{type_sum.Nd}
\alias{type_sum.Nd}
\title{Implementation of [pillar.type_sum]}
\usage{
type_sum.Nd(x)
}
\arguments{
\item{x}{object of class [`Nd`].}
}
\description{
Changes type summary (shown in tibbles) from `Surv` to `Nd`.
}
\seealso{
Other `Nd` functions: 
\code{\link{cens_ratio}()},
\code{\link{geom_Nd}()},
\code{\link{impute.Nd}()},
\code{\link{is.Nd}()}
}
\concept{`Nd` functions}
