% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{cens_ratio}
\alias{cens_ratio}
\title{Compute Degree of Censoring}
\usage{
cens_ratio(x, na.rm = FALSE)
}
\arguments{
\item{x}{object of class [`Nd`].}

\item{na.rm}{logical. NA observations ignored when `TRUE`.}
}
\description{
Compute Degree of Censoring
}
\seealso{
Other `Nd` functions: 
\code{\link{geom_Nd}()},
\code{\link{impute.Nd}()},
\code{\link{is.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{`Nd` functions}
