% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{geom_Nd}
\alias{geom_Nd}
\title{Visualise Non-Detect Data}
\usage{
geom_Nd(x)
}
\arguments{
\item{x}{object of class \code{\link{Nd}}.}
}
\description{
Visualise Non-Detect Data
}
\seealso{
Other \code{Nd} functions: 
\code{\link{cens_ratio}()},
\code{\link{impute.Nd}()},
\code{\link{is.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{\code{Nd} functions}
