% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nd-methods.R
\name{is.Nd}
\alias{is.Nd}
\title{Check \code{Nd} Class}
\usage{
is.Nd(x)
}
\arguments{
\item{x}{object of class \code{\link{Nd}}.}
}
\description{
Check \code{Nd} Class
}
\seealso{
Other \code{Nd} functions: 
\code{\link{cens_ratio}()},
\code{\link{impute.Nd}()},
\code{\link{plot.Nd}()},
\code{\link{type_sum.Nd}()}
}
\concept{\code{Nd} functions}
