% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-censreg-data.R
\name{simulate_censreg_data}
\alias{simulate_censreg_data}
\title{Simulate Censored Regression Data}
\usage{
simulate_censreg_data(n, coeff, noise, quantiles)
}
\arguments{
\item{n}{intger. Number of observations.}

\item{regr}{vector. Regression coefficients, including intercept.}
}
\description{
Simulates linear regression and [artificially censors][artificial_censor()]
  the dependent variable.
}
\note{
The number of independent variabels
}
